Move student_attribute_proxy into student_application module
